Rename efpFactory to eFP.Viewer.util
